refactor: move mobi conversion code to convert package
