+++ convert/starDict.go
@@ -24,7 +24,8 @@ idxfilesize={{.Idxfilesize}}
 sametypesequence=h
 description=Dictionary generated by runik https://runik.app`
 
-// TODO-MM: move other convert funcs to this package
+// Convert to the stardict dictionary format
+// https://stardict-4.sourceforge.net/StarDictFileFormat
 func StarDict(dictPath string, outDir string, appConfigDir string) error {
 	dictionary, err := core.DictFromFile(dictPath)
 	if err != nil {
